omit management commands from coverage,
