+++ authentik/blueprints/tests/test_bundled.py
@@ -6,11 +6,18 @@ from typing import Callable
 from django.test import TransactionTestCase
 from django.utils.text import slugify
 
+from authentik.blueprints.tests import apply_blueprint
 from authentik.blueprints.v1.importer import Importer
+from authentik.tenants.models import Tenant
 
 
 class TestBundled(TransactionTestCase):
     """Empty class, test methods are added dynamically"""
+
+    @apply_blueprint("blueprints/default/90-default-tenant.yaml")
+    def test_decorator_static(self):
+        """Test @apply_blueprint decorator"""
+        self.assertTrue(Tenant.objects.filter(domain="authentik-default").exists())
 
 
 def blueprint_tester(file_name: str) -> Callable:
