+++ authentik/core/management/commands/bootstrap_tasks.py
@@ -4,7 +4,7 @@ from django.core.management.base import BaseCommand
 from authentik.root.celery import _get_startup_tasks
 
 
-class Command(BaseCommand):  # pragma: no cover
+class Command(BaseCommand):
     """Run bootstrap tasks to ensure certain objects are created"""
 
     def handle(self, **options):
